Error if package already exists.

+++ scripts/make-package.rb
@@ -10,6 +10,11 @@ end
 version = JSON.load(File.read('src/manifest.json'))['version']
 out = "marinara-#{version}.zip"
 
+if File.exist?(out)
+  puts "#{out} already exists."
+  exit 1
+end
+
 Zip::ZipFile::open(out, 'w') do |zip|
   Dir['src/**/*'].each do |path|
     zip.add(path.sub(/^src\//, ''), path)
